Move namespace of ScoreToken classes
