+++ app/Models/Multiplayer/Room.php
@@ -10,9 +10,9 @@ use App\Exceptions\InvariantException;
 use App\Models\Beatmap;
 use App\Models\Chat\Channel;
 use App\Models\Model;
+use App\Models\ScoreToken;
 use App\Models\Season;
 use App\Models\SeasonRoom;
-use App\Models\Solo\ScoreToken;
 use App\Models\Traits\WithDbCursorHelper;
 use App\Models\User;
 use App\Traits\Memoizes;
