+++ app/Models/Solo/Score.php
@@ -13,6 +13,7 @@ use App\Models\Beatmap;
 use App\Models\Model;
 use App\Models\Multiplayer\ScoreLink as MultiplayerScoreLink;
 use App\Models\Score as LegacyScore;
+use App\Models\ScoreToken;
 use App\Models\Traits;
 use App\Models\User;
 use Carbon\Carbon;
