+++ app/Transformers/ScoreTokenTransformer.php
@@ -3,10 +3,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Transformers\Solo;
+namespace App\Transformers;
 
-use App\Models\Solo\ScoreToken;
-use App\Transformers\TransformerAbstract;
+use App\Models\ScoreToken;
 
 class ScoreTokenTransformer extends TransformerAbstract
 {
