+++ tests/api_routes.json
@@ -116,7 +116,7 @@
         "methods": [
             "POST"
         ],
-        "controller": "App\\Http\\Controllers\\Solo\\ScoreTokensController@store",
+        "controller": "App\\Http\\Controllers\\ScoreTokensController@store",
         "middlewares": [
             "App\\Http\\Middleware\\ThrottleRequests:1200,1,api:",
             "App\\Http\\Middleware\\RequireScopes",
